changes made to domain exceptions and added them to appropriate directories, DTOs added to http layer, .
