+++ src/APP/Application/ports/IEncryptionService.ts
