+++ src/APP/Domain/entities/User.ts
@@ -1,29 +1,30 @@
-import { InvalidAgeException } from "../exceptions/user";
+import { InvalidAgeException } from "../exceptions/user/InvalidAge.exception";
 import { UserAttributes } from "../types/user";
-import { Credentials } from "../valueObjects/user/Credentials";
 
 class User {
     userId: string;
     name: string;
     email: string;
-    credentials: Credentials;
+    userName: string;
+    password: string;
     age: number;
 
-    constructor(userId: string = "NULL", name: string = "NULL", userName: string = "NULL", email: string = "null@null.com", password: string = "null", age: number = 1) {
+    constructor(userId: string, name: string, userName: string, email: string, password: string, age: number) {
         this.enforceMinAgeLimit(age);
         this.userId = userId;
         this.name = name;
-        this.credentials = Credentials.create(userName, password);
+        this.userName = userName;
+        this.password = password;
         this.email = email;
         this.age = age;
     }
 
     get _userName() {
-        return this.credentials._username;
+        return this.userName;
     }
 
     get _password() {
-        return this.credentials._password;
+        return this.password;
     }
 
     static createByParams(userId: string, name: string, userName: string, email: string, password: string, age: number): User {
