+++ src/APP/Domain/exceptions/pagination/InsufficientPaginatedRows.exception.ts
@@ -1,0 +1,7 @@
+import { BaseException } from "../BaseException";
+
+export class InsufficientPaginatedRowsException extends BaseException {
+    constructor(message: string) {
+        super(message);
+    }
+}
