+++ src/APP/Domain/exceptions/pagination/PaginationOptions.exception.ts
@@ -1,0 +1,7 @@
+import { BaseException } from "../BaseException";
+
+export class PaginationOptionsException extends BaseException {
+    constructor(message: string) {
+        super(message);
+    }
+}
