--- src/APP/Domain/exceptions/todo.ts
@@ -1,7 +1,0 @@
-import { BaseException } from "./BaseException";
-
-export class InvalidTodoIdException extends BaseException {
-    constructor(message: string) {
-        super(message);
-    }
-}
