+++ src/APP/Domain/exceptions/user/InvalidAge.exception.ts
@@ -1,0 +1,7 @@
+import { BaseException } from "../BaseException";
+
+export class InvalidAgeException extends BaseException {
+    constructor() {
+        super("Invalid age, age should be greater than 5.");
+    }
+}
