+++ src/APP/Infrastructure/adapters/services/UUIDGenerator.ts
@@ -1,5 +1,5 @@
 import { v4 as uuid } from "uuid";
-import { IUniqueIDGenerator } from "../../Application/contracts/IUniqueIDGenerator";
+import { IUniqueIDGenerator } from "../../../Application/ports/IUniqueIDGenerator";
 
 export class UUIDGenerator implements IUniqueIDGenerator {
     constructor() { }
