+++ src/HTTP/DTO/todo.dto.ts
@@ -1,4 +1,4 @@
-import Todo from "../../Domain/entities/Todo";
+import Todo from "../../APP/Domain/entities/Todo";
 
 export class TodoDTO {
   todoId: string;
@@ -21,17 +21,5 @@ export class TodoDTO {
       completed: todo.completed,
     }
   }
-
-  public static toPersistence(todo: Todo): TodoDTO {
-    return {
-      todoId: todo.todoId,
-      title: todo.title,
-      description: todo.description,
-      completed: todo.completed,
-    }
-  }
-
-  public static toDomain(todo: TodoDTO): Todo {
-    return Todo.createByObject(todo);
-  }
+  
 }
