+++ src/HTTP/bootstrap/app.ts
@@ -10,7 +10,6 @@ const app = express();
 app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-
 addRoutes(app);
 
 export default app;
